requirements.txt
# FastAPI and Server
fastapi==0.127.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
sqlalchemy>=2.0.0

# Celery and Redis
celery==5.6.0
redis==7.1.0
kombu==5.6.1

# Data Validation
pydantic==2.12.4
pydantic-settings==2.12.0

# Async Support
aiofiles==25.1.0
httpx==0.28.1
aiohttp==3.13.2

# CORS
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.2.1

# Tencent Cloud SDK
tencentcloud-sdk-python==3.1.11
